requirements.txt
fastapi
openai
python-dotenv
phospho
replicate
uvicorn
pinecone-client
python-dotenv
scikit-learn
groq